python - create rdf script
